Separate upload video and generate playlist functions into two modules
